Front/back/marlisson/buildingpages3

+++ backend/app/router/company_router.py
@@ -1,12 +1,11 @@
-from fastapi import APIRouter, Form, File, UploadFile, status, Depends, Request
-from app.repository.company.models.company_models import PutCompany, GetCompanyResponse, RegisterCompanyResponse, GetCompanyData
+from app.repository.company.models.company_models import GetCompanyResponse, RegisterCompanyResponse
 from app.repository.company.sqlalchemy import CompanyRepository
 from app.repository.company.controller import CompanyController
 from app.repository.company.service import CompanyService
 from app.utils.auth import get_authenticated_user
+from fastapi import APIRouter, status, Form, Depends, Request
 from typing import Any, Annotated, Optional
 from sqlalchemy.orm import Session
-import logging
 
 repository = CompanyRepository()
 service = CompanyService(repository)
@@ -19,14 +18,7 @@ def get_db(request: Request):
 
 @router.get('/{company_id}', status_code=status.HTTP_200_OK, response_model=GetCompanyResponse)
 def get_company_by_id(company_id: str, current_user: dict = Depends(get_authenticated_user), session: Session = Depends(get_db)) -> Any:
-    if current_user['role'] == 'admin':
-        return controller.get_company_by_id_controller(company_id, session)
-    else:
-        return GetCompanyResponse(
-            status=False,
-            message='Acesso negado.',
-            data=None,
-        )
+    return controller.get_company_by_id_controller(company_id, session)
 
 
 @router.get('/', status_code=status.HTTP_200_OK, response_model=GetCompanyResponse)
@@ -47,15 +39,23 @@ def create_company(
     area: Annotated[str, Form()],
     description: Annotated[str, Form()],
     localization: Annotated[str, Form()],
+    current_user: dict = Depends(get_authenticated_user),
     session: Session = Depends(get_db)
 ) -> Any:
-    return controller.create_company_controller(
-        name,
-        area,
-        description,
-        localization,
-        session
-    )
+    if current_user['role'] == 'admin':
+        return controller.create_company_controller(
+            name,
+            area,
+            description,
+            localization,
+            session
+        )
+    else:
+        return RegisterCompanyResponse(
+            status=False,
+            message='Acesso negado.',
+            data=None,
+        )
 
 @router.put('/{company_id}', status_code=status.HTTP_200_OK, response_model=GetCompanyResponse)
 def update_company(
@@ -67,9 +67,29 @@ def update_company(
     current_user: dict = Depends(get_authenticated_user),
     session: Session = Depends(get_db)
 ) -> Any:
-    company_data = PutCompany(name=name, area=area, description=description, localization=localization)
-    return controller.update_company_controller(company_id, company_data, session)
+    if current_user['role'] == 'admin':
+        return controller.update_company_controller(
+            company_id, 
+            name,
+            area,
+            description,
+            localization,
+            session
+        )
+    else:
+        return GetCompanyResponse(
+            status=False,
+            message='Acesso negado.',
+            data=None,
+        )
 
-@router.delete('/{company_id}', status_code=status.HTTP_204_NO_CONTENT)
+@router.delete('/{company_id}', status_code=status.HTTP_200_OK, response_model=GetCompanyResponse)
 def delete_company(company_id: str, current_user: dict = Depends(get_authenticated_user), session: Session = Depends(get_db)) -> None:
-    controller.delete_company_controller(company_id, session)
+    if current_user['role'] == 'admin':
+        return controller.delete_company_controller(company_id, session)
+    else:
+        return GetCompanyResponse(
+            status=False,
+            message='Acesso negado.',
+            data=None,
+        )
